% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_cmd.R
\name{clean_restart}
\alias{clean_restart}
\title{Reboot rsession cleanly}
\usage{
clean_restart()
}
\description{
Ensure that packages are detached correctly by removing objects before doing
a rsession restart
}
