% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_cmd.R
\name{send_cmd}
\alias{send_cmd}
\title{Apply function on highlighted text in console}
\usage{
send_cmd(fn)
}
\arguments{
\item{fn}{name of the function to call}
}
\description{
Generic function to run function (fn) with highlighted text as input
}

