% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_cmd.R
\name{send_fc_arg}
\alias{send_fc_arg}
\title{Run function's parameters highlited as R commands}
\usage{
send_fc_arg(str_arg)
}
\arguments{
\item{str_arg}{String of the arguments}
}
\description{
Replace the ',' in the function call with ';' so it's understood by R
}
