% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_cmd.R
\name{send_str}
\alias{send_str}
\title{Send str to R console}
\usage{
1) Highlight variable or code by double clicking or selecting
2) Click the Addins menu
3) Select 'Call str'

Alternatively, you can use shortcut keys
}
\description{
Call str in the R console using the highlighted text as the input
}
\note{
Function depends on 1) RStudio (>= 0.99.878) with support for addins,
and 2) getActiveDocumentContext & sendToConsole from
rstudioapi package (>= 0.5-1)
}

